Idea to improv
I just start to use here in my team Sketchpackage and I really like. It is saving time and confusion.
Because then it is not necessary to every designer install everyplugin all the time.

I have a small idea for you.
When we talk about teams, there are two sort of plugins,
1. the plugins that are interesting for the whole team and 2. the plugins that only a designer uses in his workflow.
Some designers use to use a huge number of plugins and some of them just get confusing with them and use few.

It will be amazing if all the plugins from Sketchpackage could be clusted in a folder inside the plugins structure.

example:  I open my sketch, go to:    Plugins/Sketchpackage/Renameit    
Plugins ( plugins painel )
Sketchpackage ( a sublevel of navegation with all the sketchpackage plugins )
Plugin ( the plugin itself, "Renameit" for example )

This will be pretty useful to know what plugins are global in the middle of the list.
Futher more, for the future, it will be AMAZING if will be possible to do your own clusters.
Example: I can create a cluster called "Portal team" and other called "Mobile team" or  "payments team" 
or by cliente "Cliente1" . "Cliente2".  But this is a more complex thing i guess.

I just thing if you are helping people to organize plugins, this will be probably the next step.

Thanks for your time and attetion.
Best,
Murilo

murilo.souzaf@gmail.com